Set up unit tests for PipelineSchedules (#944)
- Moved e2e tests to a different file
- Set up unit tests to test smaller pieces (currently testing `init` and
`forward` for varying inputs/outputs)
- Fix bugs in PipelineStage which prevent e2e from running
